moved the rest of btn's code from loaned.js to a new components
