+++ client/src/pages/Loaned.js
@@ -22,43 +22,24 @@ import PerfectScrollbar from "perfect-scrollbar";
 // reactstrap components
 import {
 
-  Card,
-  CardHeader,
-  CardBody,
-  CardFooter,
-  CardTitle,
-  ListGroupItem,
-  ListGroup,
+ 
   Container,
   Row,
   Col,
   Button,
-  Label,
-  FormGroup,
-  Form,
-  Input,
-  FormText,
-  NavItem,
-  NavLink,
-  Nav,
-  Table,
-  TabContent,
-  TabPane,
-  UncontrolledTooltip,
-  UncontrolledCarousel,
-  Modal,
-  InputGroup,
-  InputGroupAddon,
-  InputGroupText,
+  UncontrolledCarousel
 } from "reactstrap";
 
 // core components
 import ToolShedNavbar from "components/Navbars/ToolShedNavbar.js";
 import AddBtnLoaned from "components/AddBtnLoaned/AddBtnLoaned.js";
+import SearchBtnLoaned from "components/SearchBtnLoaned/SearchBtnLoaned.js";
+import SearchOutNetwork from "components/SearchOutNetwork/SearchOutNetwork.js";
+import UserNameDisplay from "components/UserNameDisplay/UserNameDisplay.js";
+import ExampleToolList from "components/ExampleToolList/ExampleToolList.js";
 import Footer from "components/Footer/Footer.js";
 import ToolCard from "components/ToolCard/Toolcard.js";
 // ====================== post import
-import { Inputs, TextArea, FormBtn } from "../components/Form";
 import API from 'utils/toolshed-api';
 import { useEffect, useState } from "react";
 
@@ -76,9 +57,7 @@ const carouselItems = [
 let ps = null;
 
 export default function Loaned() {
- 
-  const [formModalSearch, setFormModalSearch] = React.useState(false);
-  const [tabs, setTabs] = React.useState(1);
+
   React.useEffect(() => {
     if (navigator.platform.indexOf("Win") > -1) {
       document.documentElement.className += " perfect-scrollbar-on";
@@ -102,21 +81,7 @@ export default function Loaned() {
     loadTools()
   }, [])
 
-   // Handles updating component state when the user types into the input field
-   function handleInputChange(event) {
-    const { name, value } = event.target;
-    setFormObject({ ...formObject, [name]: value })
-  };
-
   
-
-  const [formObject, setFormObject] = useState({
-    name: "",
-    description: "",
-    price: [],
-    categoryId: []
-  });
-
   return (
     <>
       <ToolShedNavbar />
@@ -162,146 +127,15 @@ export default function Loaned() {
                   <div className="btn-wrapper pt-3">
 
                     <div className="button-container">
-                 
-                 <AddBtnLoaned/>
-                    
-                    
-                      <Button
-                        className="btn-icon btn-simple btn-round btn-neutral"
-                        color="success" id="tooltip21" onClick={() => setFormModalSearch(true)}>
-                        <i className="tim-icons icon-zoom-split" />
-                      </Button>
-                      <UncontrolledTooltip delay={0} placement="right" target="tooltip21">
-                        Search Loaned Tools
-                      </UncontrolledTooltip>
-                      {/* Start Search Form Modal */}
-                      <Modal
-                        modalClassName="modal-black"
-                        isOpen={formModalSearch}
-                        toggle={() => setFormModalSearch(false)}
-                      >
-                        <div className="modal-header justify-content-center">
-                          <button className="close" onClick={() => setFormModalSearch(false)}>
-                            <i className="tim-icons icon-simple-remove text-white" />
-                          </button>
-                          <div className="text-muted text-center ml-auto mr-auto">
-                            <h3 className="mb-0">Search Your Loaned Tools</h3>
-                          </div>
-                        </div>
-                        <div className="modal-body">
-                          <div className="btn-wrapper text-center">
 
-                          </div>
-                          <div className="text-center text-muted mb-4 mt-3">
-                            <small>Describe what you are looking for in this directory</small>
-                          </div>
-                          <Form role="addForm">
-                            <FormGroup className="mb-3">
-                              <InputGroup>
-                                <InputGroupAddon addonType="prepend">
-                                  <InputGroupText>
-                                    Name:
-                                  </InputGroupText>
-                                </InputGroupAddon>
-                                <Input
-                                  onChange={handleInputChange}
-                                  name="name"
-                                  placeholder="name (required)"
-                                  value={formObject.name}
-                                />
-                              </InputGroup>
-                            </FormGroup>
-                            <FormGroup>
-                              <InputGroup>
-                                <InputGroupAddon addonType="prepend">
-                                  <InputGroupText>
-                                    Category:
-                                </InputGroupText>
-                                </InputGroupAddon>
-                                <Input
-                                  placeholder="Tool Category"
-                                  type="text"
-                                />
-                              </InputGroup>
-                            </FormGroup>
-                            <FormGroup>
-                              <InputGroup>
-                                <InputGroupAddon addonType="prepend">
-                                  <InputGroupText>
-                                    Description:
-                                  </InputGroupText>
-                                </InputGroupAddon>
-                                <Input
-                                  placeholder="Tool Description"
-                                  type="text"
-                                />
-                              </InputGroup>
-                            </FormGroup>
-                            <FormGroup>
-                              <InputGroup>
-                                <InputGroupAddon addonType="prepend">
-                                  <InputGroupText>
-                                    Price:
-                                  </InputGroupText>
-                                </InputGroupAddon>
-                                <Input
-                                  placeholder="Tool Price"
-                                  type="text"
-                                />
-                              </InputGroup>
-                            </FormGroup>
-                            <FormGroup>
-                              <InputGroup>
-                                <InputGroupAddon addonType="prepend">
-                                  <InputGroupText>
-                                    Loaned To:
-                                  </InputGroupText>
-                                </InputGroupAddon>
-                                <Input
-                                  placeholder="User Name"
-                                  type="text"
-                                />
-                              </InputGroup>
-                            </FormGroup>
-                            <FormGroup>
-                              <InputGroup>
-                                <InputGroupAddon addonType="prepend">
-                                  <InputGroupText>
-                                    Promise Date:
-                                  </InputGroupText>
-                                </InputGroupAddon>
-                                <Input
-                                  placeholder="Expected Return Date"
-                                  type="text"
-                                />
-                              </InputGroup>
-                            </FormGroup>
+                      <AddBtnLoaned />
 
 
-                            <div className="text-center">
-                              <Button className="my-4" color="primary" type="button">
-                                Search TOOL
-                              </Button>
-                            </div>
-                          </Form>
-                        </div>
-                      </Modal>
-                      {/* End Search Form Modal */}
+                      <SearchBtnLoaned />
                     </div>
 
                     <div className="button-container">
-                      <Button
-                        className="btn-simple"
-                        color="info"
-                        href="#pablo"
-                        id="tooltip22"
-                        onClick={(e) => e.preventDefault()}
-                      >
-                        <i className="tim-icons icon-bulb-63" /> Search It!
-                      </Button>
-                      <UncontrolledTooltip delay={0} placement="bottom" target="tooltip22">
-                        Search Outside of Directory
-                      </UncontrolledTooltip>
+                      <SearchOutNetwork />
                     </div>
 
 
@@ -311,12 +145,7 @@ export default function Loaned() {
 
               </Row>
 
-              <Row>
-                <Col lg="6" md="6">
-                  <h2 className="profile-title text-left">Michelle Fairbanks</h2>
-                  <h4 className="text-on-back">UserName</h4>
-                </Col>
-              </Row>
+             <UserNameDisplay/>
 
             </Container>
 
@@ -331,7 +160,7 @@ export default function Loaned() {
 
         </div>
 
-        {/* Loaned Tools list section */}
+
         <section className="section section-lg section-coins">
           <img
             alt="..."
@@ -344,120 +173,13 @@ export default function Loaned() {
                 <hr className="line-info" />
                 <h1>
                   Loaned Out{" "}
-                  <span className="text-info">please return as received</span>
+                  <span className="text-info">Please return as received</span>
                 </h1>
               </Col>
             </Row>
-            <Row>
-              <Col md="4">
-                <Card className="card-coin card-plain">
-                  <CardHeader>
-                    <img
-                      alt="..."
-                      className="img-center img-fluid"
-                      src={require("assets/img/mechanic-toolset.jpg").default}
-                    />
-                  </CardHeader>
-                  <CardBody>
-                    <Row>
-                      <Col className="text-center" md="12">
-                        <h4 className="text-uppercase">SAE Wrench set</h4>
-                        <span>Hand Tools</span>
-                        <hr className="line-primary" />
-                      </Col>
-                    </Row>
-                    <Row>
-                      <ListGroup>
-                        {/* <ListGroupItem>Category: {{category}}</ListGroupItem> */}
-                        <ListGroupItem>Description</ListGroupItem>
-                        <ListGroupItem>Date Loaned Out</ListGroupItem>
-                        <ListGroupItem>Return Promise</ListGroupItem>
-                      </ListGroup>
-                    </Row>
-                  </CardBody>
-                  <CardFooter className="text-center">
-                    <Button className="btn-simple" color="primary">
-                      Edit
-                    </Button>
-                    <Button className="btn-simple" color="danger">
-                      Remove
-                    </Button>
-                  </CardFooter>
-                </Card>
-              </Col>
-              <Col md="4">
-                <Card className="card-coin card-plain">
-                  <CardHeader>
-                    <img
-                      alt="..."
-                      className="img-center img-fluid"
-                      src={require("assets/img/protective-gear.jpg").default}
-                    />
-                  </CardHeader>
-                  <CardBody>
-                    <Row>
-                      <Col className="text-center" md="12">
-                        <h4 className="text-uppercase">Welder Gloves</h4>
-                        <span>Protective Gear</span>
-                        <hr className="line-success" />
-                      </Col>
-                    </Row>
-                    <Row>
-                      <ListGroup>
-                        {/* <ListGroupItem>Category: {{category}}</ListGroupItem> */}
-                        <ListGroupItem>Description</ListGroupItem>
-                        <ListGroupItem>Date Loaned Out</ListGroupItem>
-                        <ListGroupItem>Return Promise</ListGroupItem>
-                      </ListGroup>
-                    </Row>
-                  </CardBody>
-                  <CardFooter className="text-center">
-                    <Button className="btn-simple" color="success">
-                      Edit
-                    </Button>
-                    <Button className="btn-simple" color="danger">
-                      Remove
-                    </Button>
-                  </CardFooter>
-                </Card>
-              </Col>
-              <Col md="4">
-                <Card className="card-coin card-plain">
-                  <CardHeader>
-                    <img
-                      alt="..."
-                      className="img-center img-fluid"
-                      src={require("assets/img/power-tools.jpg").default}
-                    />
-                  </CardHeader>
-                  <CardBody>
-                    <Row>
-                      <Col className="text-center" md="12">
-                        <h4 className="text-uppercase">Impact Wrench</h4>
-                        <span>Power Tools</span>
-                        <hr className="line-info" />
-                      </Col>
-                    </Row>
-                    <Row>
-                      <ListGroup>
-                        {/* <ListGroupItem>Category: {{category}}</ListGroupItem> */}
-                        <ListGroupItem>Description</ListGroupItem>
-                        <ListGroupItem>Date Loaned Out</ListGroupItem>
-                        <ListGroupItem>Return Promise</ListGroupItem>
-                      </ListGroup>
-                    </Row>
-                  </CardBody>
-                  <CardFooter className="text-center">
-                    <Button className="btn-simple" color="info">
-                      Edit
-                    </Button>
-                    <Button className="btn-simple" color="danger">
-                      Remove
-                    </Button>
-                  </CardFooter>
-                </Card>
-              </Col>
-            </Row>
+            {/* Loaned Tools list section */}
+            <ExampleToolList />
+
           </Container>
         </section>
